Added US03 Test and function to generate an individual object
Completed US03 testing. Instead of adding to the GEDCOM file to test each case, I created a function to easily an individual object which can be used to test:

Created the gen_individual function in the unit test file.
Parameters: gen_individual(self, name, sex, birt, deat)
Instead of adding to the GEDCOM file to test each case, I created this function to easily generate an individual object which can be used to test. It doesn't take family into consideration.

All unit tests are working and no longer interfere with eachother.
I will work on creating a function to generate a family and the unit test
for US04 tomorrow.